# be/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: giacomo.cortesi1993@gmail.com
    name: Giacomo Cortesi
    url: https://giacomocortesi.github.io/hugo-gc
  description: AI Music to Subtitles API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AI Music Subtitles - OpenAPI 3.0
  version: 1.0.0
servers:
- description: API server
  url: "http://{username}:{port}"
  variables:
    username:
      default: localhost
    port:
      default: "8080"
      enum:
      - "8080"
      - "443"
tags:
- description: Music AI subtitles generator jobs
  name: job
- description: Transcription CRUD operations
  name: transcription
- description: Audio/Video files CRUD operations
  name: file
paths:
  /file:
    delete:
      operationId: file_delete
      parameters:
      - description: "Filename of the audio or video file to be deleted, if no filename,\
          \ all uploaded files are deleted"
        explode: true
        in: query
        name: filename
        required: false
        schema:
          type: string
        style: form
      responses:
        "204":
          description: File successfully deleted
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Delete uploaded file(s)
      tags:
      - file
      x-openapi-router-controller: openapi_server.controllers.file_controller
    get:
      operationId: file_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FileResponse'
                type: array
          description: Files successfully retrieved
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Retrieves the list of uploaded files
      tags:
      - file
      x-openapi-router-controller: openapi_server.controllers.file_controller
    post:
      operationId: file_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/_file_post_request'
        required: true
      responses:
        "200":
          description: File successfully uploaded
        "400":
          content:
            application/problem+json:
              example:
                title: BadRequest
                detail: Filename missing'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Uploads a video or audio file
      tags:
      - file
      x-openapi-router-controller: openapi_server.controllers.file_controller
  /job:
    get:
      operationId: job_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/JobResponse'
                type: array
          description: Jobs details successfully retrieved
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Get all job details
      tags:
      - job
      x-openapi-router-controller: openapi_server.controllers.job_controller
    post:
      operationId: job_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
          description: Subtitles generation job successfully created
        "400":
          content:
            application/problem+json:
              example:
                title: Bad Request
                detail: '''test'' is not of type ''boolean'' - ''config.speaker_detection'''
                status: 400
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Create a new subtitles generation job
      tags:
      - job
      x-openapi-router-controller: openapi_server.controllers.job_controller
  /job/{id}:
    get:
      operationId: job_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
          description: Job details successfully retrieved
        "404":
          content:
            application/problem+json:
              example:
                title: NotFound
                detail: The requested job ID was not found on the server
                status: 404
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Get job details
      tags:
      - job
      x-openapi-router-controller: openapi_server.controllers.job_controller
  /transcription:
    get:
      operationId: transcription_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TranscriptionResponse'
                type: array
          description: Transcriptions successfully retrieved
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Fetch all transcriptions
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
    post:
      operationId: transcription_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_transcription_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionResponse'
          description: Transcription succesfully created
        "400":
          content:
            application/problem+json:
              example:
                title: Bad Request
                detail: Bad Request
                status: 400
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Creates a new transcription
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
  /transcription/{id}:
    delete:
      operationId: transcription_id_delete
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Transcription successfully deleted
        "404":
          content:
            application/problem+json:
              example:
                title: NotFound
                detail: The requested transcription ID was not found on the server
                status: 404
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Deletes a specific transcription
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
    get:
      operationId: transcription_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionResponse'
          description: Transcription successfully retrieved
        "404":
          content:
            application/problem+json:
              example:
                title: NotFound
                detail: The requested transcription ID was not found on the server
                status: 404
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Retrieves a specific transcription
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
    patch:
      operationId: transcription_id_patch
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_transcription__id__patch_request'
        required: true
      responses:
        "204":
          description: Transcriptions successfully edited
        "404":
          content:
            application/problem+json:
              example:
                title: NotFound
                detail: The requested transcription ID was not found on the server
                status: 404
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Edit a specific transcription
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
  /transcription/{id}/clear:
    post:
      operationId: transcription_id_clear_post
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Transcription successfully cleared
        "404":
          content:
            application/problem+json:
              example:
                title: NotFound
                detail: The requested transcription ID was not found on the server
                status: 404
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Restores initial transcription
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
  /transcription/{id}/export:
    get:
      operationId: transcription_id_export_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: The format to export the transcription in
        explode: true
        in: query
        name: format
        required: false
        schema:
          enum:
          - srt
          - stt
          - video
          type: string
        style: form
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Successful export
        "400":
          content:
            application/problem+json:
              example:
                title: BadRequest
                detail: Unsupported export format
                status: 400
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unsupported format
        "404":
          content:
            application/problem+json:
              example:
                title: NotFound
                detail: The requested transcription ID was not found on the server
                status: 404
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Export transcription in several subtitle formats
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
  /transcription/{id}/fit:
    post:
      operationId: transcription_id_fit_post
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Transcription fit action succesfull
        "404":
          content:
            application/problem+json:
              example:
                title: NotFound
                detail: The requested transcription ID was not found on the server
                status: 404
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Fit start and end of each subtitles segment
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
  /transcription/{id}/fix:
    post:
      operationId: transcription_id_fix_post
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Transcription fix action succesfull
        "404":
          content:
            application/problem+json:
              example:
                title: NotFound
                detail: The requested transcription ID was not found on the server
                status: 404
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        "500":
          content:
            application/problem+json:
              example:
                title: Internal Server Error
                detail: Unexpected error'
                status: 500
                type: about:blank
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Internal Server Error
      summary: Attempts to fix all subtitles text with AI
      tags:
      - transcription
      x-openapi-router-controller: openapi_server.controllers.transcription_controller
components:
  schemas:
    JobInfo:
      example:
        filename: filename
        config:
          speaker_detection: true
          subtitles_frequency: 0
          language: language
          model_size: model_size
      properties:
        config:
          $ref: '#/components/schemas/JobConfig'
        filename:
          title: filename
          type: string
      title: JobInfo
      type: object
    JobConfig:
      example:
        speaker_detection: true
        subtitles_frequency: 0
        language: language
        model_size: model_size
      properties:
        model_size:
          title: model_size
          type: string
        subtitles_frequency:
          title: subtitles_frequency
          type: integer
        language:
          title: language
          type: string
        speaker_detection:
          title: speaker_detection
          type: boolean
      required:
      - language
      - model_size
      - speaker_detection
      - subtitles_frequency
      title: JobConfig
      type: object
    JobRequest:
      example:
        filename: filename
        config:
          speaker_detection: true
          subtitles_frequency: 0
          language: language
          model_size: model_size
      properties:
        filename:
          title: filename
          type: string
        config:
          $ref: '#/components/schemas/JobConfig'
      required:
      - filename
      title: JobRequest
      type: object
    JobResponse:
      example:
        data:
          language: language
          segments:
          - start: 6.0274563
            words:
            - score: 2.3021358869347655
              start: 5.962133916683182
              end: 5.637376656633329
              word: word
            - score: 2.3021358869347655
              start: 5.962133916683182
              end: 5.637376656633329
              word: word
            end: 1.4658129
            text: text
          - start: 6.0274563
            words:
            - score: 2.3021358869347655
              start: 5.962133916683182
              end: 5.637376656633329
              word: word
            - score: 2.3021358869347655
              start: 5.962133916683182
              end: 5.637376656633329
              word: word
            end: 1.4658129
            text: text
          word_segments:
          - score: 2.3021358869347655
            start: 5.962133916683182
            end: 5.637376656633329
            word: word
          - score: 2.3021358869347655
            start: 5.962133916683182
            end: 5.637376656633329
            word: word
        id: id
        info:
          filename: filename
          config:
            speaker_detection: true
            subtitles_frequency: 0
            language: language
            model_size: model_size
        status: pending
      properties:
        id:
          title: id
          type: string
        info:
          $ref: '#/components/schemas/JobInfo'
        status:
          enum:
          - pending
          - running
          - completed
          - failed
          title: status
          type: string
        data:
          $ref: '#/components/schemas/TranscriptionData'
      required:
      - config
      - data
      - filename
      - id
      - status
      title: JobResponse
      type: object
    Word:
      example:
        score: 2.3021358869347655
        start: 5.962133916683182
        end: 5.637376656633329
        word: word
      properties:
        word:
          title: word
          type: string
        start:
          title: start
          type: number
        end:
          title: end
          type: number
        score:
          title: score
          type: number
      required:
      - end
      - score
      - start
      - word
      title: Word
      type: object
    Segment:
      example:
        start: 6.0274563
        words:
        - score: 2.3021358869347655
          start: 5.962133916683182
          end: 5.637376656633329
          word: word
        - score: 2.3021358869347655
          start: 5.962133916683182
          end: 5.637376656633329
          word: word
        end: 1.4658129
        text: text
      properties:
        start:
          format: float
          title: start
          type: number
        end:
          format: float
          title: end
          type: number
        text:
          title: text
          type: string
        words:
          items:
            $ref: '#/components/schemas/Word'
          title: words
          type: array
      required:
      - end
      - start
      - text
      title: Segment
      type: object
    TranscriptionRequest:
      properties:
        data:
          $ref: '#/components/schemas/TranscriptionData'
        job_id:
          title: job_id
          type: string
        filename:
          title: filename
          type: string
      title: TranscriptionRequest
      type: object
    TranscriptionResponse:
      example:
        filename: filename
        data:
          language: language
          segments:
          - start: 6.0274563
            words:
            - score: 2.3021358869347655
              start: 5.962133916683182
              end: 5.637376656633329
              word: word
            - score: 2.3021358869347655
              start: 5.962133916683182
              end: 5.637376656633329
              word: word
            end: 1.4658129
            text: text
          - start: 6.0274563
            words:
            - score: 2.3021358869347655
              start: 5.962133916683182
              end: 5.637376656633329
              word: word
            - score: 2.3021358869347655
              start: 5.962133916683182
              end: 5.637376656633329
              word: word
            end: 1.4658129
            text: text
          word_segments:
          - score: 2.3021358869347655
            start: 5.962133916683182
            end: 5.637376656633329
            word: word
          - score: 2.3021358869347655
            start: 5.962133916683182
            end: 5.637376656633329
            word: word
        job_id: job_id
        id: id
      properties:
        id:
          title: id
          type: string
        data:
          $ref: '#/components/schemas/TranscriptionData'
        job_id:
          title: job_id
          type: string
        filename:
          title: filename
          type: string
      title: TranscriptionResponse
      type: object
    TranscriptionData:
      example:
        language: language
        segments:
        - start: 6.0274563
          words:
          - score: 2.3021358869347655
            start: 5.962133916683182
            end: 5.637376656633329
            word: word
          - score: 2.3021358869347655
            start: 5.962133916683182
            end: 5.637376656633329
            word: word
          end: 1.4658129
          text: text
        - start: 6.0274563
          words:
          - score: 2.3021358869347655
            start: 5.962133916683182
            end: 5.637376656633329
            word: word
          - score: 2.3021358869347655
            start: 5.962133916683182
            end: 5.637376656633329
            word: word
          end: 1.4658129
          text: text
        word_segments:
        - score: 2.3021358869347655
          start: 5.962133916683182
          end: 5.637376656633329
          word: word
        - score: 2.3021358869347655
          start: 5.962133916683182
          end: 5.637376656633329
          word: word
      properties:
        segments:
          items:
            $ref: '#/components/schemas/Segment'
          title: segments
          type: array
        word_segments:
          items:
            $ref: '#/components/schemas/Word'
          title: word_segments
          type: array
        language:
          title: language
          type: string
      required:
      - segments
      title: TranscriptionData
      type: object
    FileResponse:
      example:
        filename: filename
        video_url: video_url
        image_url: image_url
        id: id
        upload_date: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          title: id
          type: string
        filename:
          title: filename
          type: string
        upload_date:
          format: date-time
          title: upload_date
          type: string
        video_url:
          title: video_url
          type: string
        image_url:
          title: image_url
          type: string
      title: FileResponse
      type: object
    ErrorModel:
      properties:
        title:
          title: title
          type: string
        detail:
          title: detail
          type: string
        status:
          title: status
          type: integer
        type:
          title: type
          type: string
      title: ErrorModel
      type: object
    _transcription_post_request:
      properties:
        schema:
          $ref: '#/components/schemas/TranscriptionRequest'
      title: _transcription_post_request
      type: object
    _transcription__id__patch_request:
      properties:
        data:
          $ref: '#/components/schemas/TranscriptionData'
      title: _transcription__id__patch_request
      type: object
    _file_post_request:
      properties:
        file:
          format: binary
          type: string
      type: object